Resolved CID 351533 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument mapping_to_send.size() to format specifier %u was expected to have type unsigned int but has type unsigned long.

+++ src/world/Server/LogonCommClient/LogonCommClient.cpp
@@ -359,7 +359,7 @@ void LogonCommClientSocket::HandleRequestAccountMapping(WorldPacket& recvData)
 
         uncompressed.clear();
     }
-    LogNotice("LogonCommClient : Build character mapping in %u ms. (%u)", Util::GetTimeDifferenceToNow(startTime), mapping_to_send.size());
+    LogNotice("LogonCommClient : Build character mapping in %u ms. (%u)", Util::GetTimeDifferenceToNow(startTime), static_cast<uint32_t>(mapping_to_send.size()));
 }
 
 void LogonCommClientSocket::CompressAndSend(ByteBuffer& uncompressed)
